feat: add relations between models

+++ routes/modules/records.js
@@ -9,8 +9,9 @@ const User = require('../../models/user')
 // 使用者可以依類別篩選支出紀錄
 router.get('/category/:category', (req, res) => {
   const keyCategory = req.params.category
+  const userId = req.user._id
   let totalAmount = 0
-  Record.find({ category: keyCategory })
+  Record.find({ category: keyCategory, userId })
     .lean()
     .then(records => {
       for (let i = 0; i < records.length; i++) {
@@ -22,9 +23,7 @@ router.get('/category/:category', (req, res) => {
     .then(records => {
       Category.find()
         .lean()
-        .then(category => {
-          res.render('index', { keyCategory, category, records, totalAmount })
-        })
+        .then(category => res.render('index', { keyCategory, category, records, totalAmount }))
         .catch(console.error)
     })
 })
@@ -37,53 +36,46 @@ router.get('/new', (req, res) => {
 
 // 送出新支出資料
 router.post('/', (req, res) => {
-  const { name, date, category, amount, user } = req.body
+  const userId = req.user._id
+  const { name, date, category, amount } = req.body
   Category.findOne({ name: category })
-    .then(category => {
-      const categoryId = category._id
-      return User.findOne({ name: user })
-        .then(user => {
-          const userId = user._id
-          return Record.create({ name, date, category: category.name, categoryIcon: category.icon, amount, userId, categoryId })
-            .then(() => res.redirect('/'))
-            .catch(console.error)
-        })
+    .then(data => {
+      const categoryId = data._id
+      return Record.create({ name, date, category, categoryIcon: data.icon, amount, userId, categoryId })
     })
+    .then(() => res.redirect('/'))
+    .catch(console.error)
 })
 
 // 使用者可以編輯一筆支出紀錄
 // 編輯支出表單頁面
 router.get('/:id/edit', (req, res) => {
+  const userId = req.user._id
   const recordId = req.params.id
-  Record.findOne({ _id: recordId })
+  Record.findOne({ _id: recordId, userId })
     .lean()
     .then(record => {
-      const userId = record.userId
       let date = record.date
       const year = date.getFullYear().toString()
       const month = ('0' + (date.getMonth() + 1).toString()).slice(-2)
       const day = ('0' + date.getDate().toString()).slice(-2)
       date = `${year}-${month}-${day}`
-      return User.findOne({ _id: userId })
-        .lean()
-        .then(user => {
-          res.render('edit', { record, user, date })
-        })
+      return res.render('edit', { record, date })
     })
     .catch(console.error)
 })
 
 // 送出編輯後的資料
 router.put('/:id', (req, res) => {
-  const { name, date, category, amount, user } = req.body
+  const userId = req.user._id
+  const { name, date, category, amount } = req.body
   const recordId = req.params.id
-  Record.findOne({ _id: recordId })
+  Record.findOne({ _id: recordId, userId })
     .then(record => {
       record.name = name
       record.date = date
       record.category = category
       record.amount = amount
-      record.user = user
       return Category.findOne({ name: category })
         .then(category => {
           record.categoryIcon = category.icon
@@ -91,13 +83,14 @@ router.put('/:id', (req, res) => {
         })
         .then(() => res.redirect('/'))
         .catch(console.error)
-    }) 
+    })
 })
 
 // 使用者可以刪除一筆支出紀錄
 router.delete('/:id', (req, res) => {
+  const userId = req.user._id
   const recordId = req.params.id
-  return Record.findOneAndRemove({ _id: recordId })
+  return Record.findOneAndRemove({ _id: recordId, userId })
     .then(() => res.redirect('/'))
     .catch(console.error)
 })
